测试完串口及GPIO功能，解决systick异常进中断的原因，原因是st的时钟初始化库函数里面会默认开启systick的中断，现在通过疲敝解决

+++ components/libc/microlib/sources/microlib_adapter.c
@@ -51,7 +51,7 @@ int fputc (int ch, FILE *f)
 { 
     UART_HandleTypeDef *p_uart_handler = (UART_HandleTypeDef *)__g_adapter.p_cookie;        
     while ((p_uart_handler->Instance->SR & 0x40) == 0);        /* 等待发送数据完成           */
-    p_uart_handler->Instance->DR = ch;                         /* 发送数据                   */
+    p_uart_handler->Instance->DR = (uint8_t)ch;                         /* 发送数据                   */
     return ch;
 }
